Move parsing to utils.text_utils, old utils to utils.utils (with bc imports)
