--- docs/src/parsing/porter.rst
@@ -1,9 +1,0 @@
-:mod:`parsing.porter` -- Porter Stemming Algorithm
-=========================================================
-
-.. automodule:: gensim.parsing.porter
-    :synopsis: Porter Stemming Algorithm
-    :members:
-    :inherited-members:
-    :undoc-members:
-    :show-inheritance:
