--- docs/src/parsing/preprocessing.rst
@@ -1,9 +1,0 @@
-:mod:`parsing.preprocessing` -- Functions to preprocess raw text
-================================================================
-
-.. automodule:: gensim.parsing.preprocessing
-    :synopsis: Functions to preprocess raw text
-    :members:
-    :inherited-members:
-    :undoc-members:
-    :show-inheritance:
