--- docs/src/utils.rst
@@ -1,9 +1,0 @@
-:mod:`utils` -- Various utility functions
-==========================================
-
-.. automodule:: gensim.utils
-    :synopsis: Various utility functions
-    :members:
-    :inherited-members:
-    :undoc-members:
-    :show-inheritance:
