+++ docs/src/utils/text_utils.rst
@@ -1,0 +1,9 @@
+:mod:`utils.text_utils` -- Functions to preprocess raw text
+===========================================================
+
+.. automodule:: gensim.utils.text_utils
+    :synopsis: Functions to preprocess raw text
+    :members:
+    :inherited-members:
+    :undoc-members:
+    :show-inheritance:
