+++ docs/src/utils/utils.rst
@@ -1,0 +1,9 @@
+:mod:`utils` -- Various utility functions
+==========================================
+
+.. automodule:: gensim.utils.utils
+    :synopsis: Various utility functions
+    :members:
+    :inherited-members:
+    :undoc-members:
+    :show-inheritance:
