+++ gensim/utils/__init__.py
@@ -1,0 +1,2 @@
+from .utils import *
+from . import text_utils
